python_search/requirements.txt
flet>=0.28.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
